CosmosDB 1.0 code change

+++ tests/certification/state/azure/cosmosdb/cosmosdb_test.go
@@ -133,29 +133,33 @@ func TestAzureCosmosDBStorage(t *testing.T) {
 				"partitionKey": "mypartition",
 			}
 
-			test := func(setMeta, getMeta map[string]string, expectedValue string) {
+			test := func(setMeta, getMeta map[string]string, expectedValue string, expectedErr bool) {
 				// save state, default options: strong, last-write
 				err = client.SaveState(ctx, statestore, stateKey, []byte(stateValue), setMeta)
-				assert.NoError(t, err)
+				require.NoError(t, err)
 
 				// get state
 				item, err := client.GetState(ctx, statestore, stateKey, getMeta)
-				assert.NoError(t, err)
-				assert.Equal(t, expectedValue, string(item.Value))
+				if expectedErr {
+					require.Error(t, err)
+					return
+				}
+				require.NoError(t, err)
+				require.Equal(t, expectedValue, string(item.Value))
 
 				// delete state
 				err = client.DeleteState(ctx, statestore, stateKey, setMeta)
-				assert.NoError(t, err)
+				require.NoError(t, err)
 			}
 
 			// Test	with no partition key
-			test(nil, meta1, stateValue)
+			test(nil, meta1, stateValue, false)
 
 			// Test with specific partition key
-			test(meta2, meta2, stateValue)
+			test(meta2, meta2, stateValue, false)
 
 			// Test with incorrect partition key
-			test(meta2, meta1, "")
+			test(meta2, meta1, "", true)
 
 			return nil
 		}
